Split streams into MemoryStream and FileStream
