+++ src/IO/MemoryStream.php
@@ -1,0 +1,10 @@
+<?php
+namespace GIFEndec\IO;
+
+class MemoryStream extends PhpStream
+{
+    public function __construct()
+    {
+        $this->phpStream = fopen('php://memory', 'wb+');
+    }
+}
